profile information view removed body line large code, splitted into sVs, removed whitespaces
